+++ functions/package.json
@@ -13,8 +13,12 @@
   },
   "main": "index.js",
   "dependencies": {
+    "body-parser": "^1.20.3",
+    "cors": "^2.8.5",
+    "express": "^4.21.2",
     "firebase-admin": "^12.6.0",
-    "firebase-functions": "^6.0.1"
+    "firebase-functions": "^6.0.1",
+    "nodemailer": "^6.9.16"
   },
   "devDependencies": {
     "firebase-functions-test": "^3.1.0"
