+++ public/find_id.html
@@ -185,7 +185,7 @@
     function sendIdEmail(toEmail, userId) {
         // 백엔드 API를 호출해 이메일 전송 요청
         $.ajax({
-            url: "http://localhost:8080/send-id", // 백엔드 엔드포인트
+            url: "https://api-htauzstwia-uc.a.run.app/send-id", // 백엔드 엔드포인트
             method: "POST",
             contentType: "application/json",
             data: JSON.stringify({
